refactor: extract LinkImageBubble from Editor.tsx (PR A of decomposition)
First extraction in the Editor.tsx decomposition plan. Zero behavior
changes — same component, different file.

- Move LinkImageBubble (425 lines) and the BubbleTarget interface
  to src/components/LinkImageBubble.tsx as a default export.
- Editor.tsx drops 443 lines (2555 → 2113).
- 26 new characterization tests in src/__tests__/link-image-bubble.test.ts
  covering: render gate (null target, null editor, stale element),
  link target (preview/edit/follow/remove/apply, disabled empty URL),
  image target (fields visible, aspect-lock auto-fill in both
  directions, aspect unlock stops auto-fill, center toggle, setImage
  + updateAttributes on apply, deleteSelection on remove, resize
  handle visible only for images), dismissal (Escape, outside click,
  inside click doesn't dismiss, Enter applies).

The tests target the shipped contract, not the implementation.
Any future refactor of LinkImageBubble or Editor that breaks an
observable behavior fails the tests.

Bundle size unchanged. 707 tests green (681 prior + 26 new).
Build clean.
